Create merged geodesy module, added enu2xyz (and reverse) function
